replace file to view/change style / add touch logic to the HallFameCard/ use env API_BASE_URL
